requirements.txt
# AutoAttack — pinned dependency versions
# Install with: pip install -r requirements.txt

# ── Core graph & data ───────────────────────────────────────────────────────
networkx==3.3
pandas==2.2.2
numpy==1.26.4

# ── HTTP & APIs ─────────────────────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.2

# ── Configuration & serialisation ──────────────────────────────────────────
PyYAML==6.0.2

# ── Metasploit RPC ──────────────────────────────────────────────────────────
pymetasploit3==1.0.3

# ── OpenAI LLM planner ─────────────────────────────────────────────────────
openai==1.40.0

# ── Neo4j graph database ────────────────────────────────────────────────────
neo4j==5.23.1

# ── Reinforcement learning ──────────────────────────────────────────────────
gymnasium==0.29.1

# ── SSH executor ────────────────────────────────────────────────────────────
paramiko==3.4.1

# ── CLI (rich terminal UI) ──────────────────────────────────────────────────
rich==13.7.1

# ── Visualisation ───────────────────────────────────────────────────────────
plotly==5.22.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
